'秒杀上架完成'

+++ guli-coupon/src/main/java/com/mzx/gulimall/coupon/entity/SmsSeckillSkuRelationEntity.java
@@ -2,15 +2,12 @@ package com.mzx.gulimall.coupon.entity;
 
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-
-import java.math.BigDecimal;
-import java.io.Serializable;
-import java.util.Date;
 import lombok.Data;
 
+import java.io.Serializable;
+import java.math.BigDecimal;
+
 /**
- * 
- * 
  * @author ZhenXinMa
  * @email 2280480546@qq.com
  * @date 2020-07-10 17:45:21
@@ -18,40 +15,40 @@ import lombok.Data;
 @Data
 @TableName("sms_seckill_sku_relation")
 public class SmsSeckillSkuRelationEntity implements Serializable {
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	/**
-	 * id
-	 */
-	@TableId
-	private Long id;
-	/**
-	 * 
-	 */
-	private Long promotionId;
-	/**
-	 * 
-	 */
-	private Long promotionSessionId;
-	/**
-	 * 
-	 */
-	private Long skuId;
-	/**
-	 * 
-	 */
-	private BigDecimal seckillPrice;
-	/**
-	 * 
-	 */
-	private BigDecimal seckillCount;
-	/**
-	 * ÿ
-	 */
-	private BigDecimal seckillLimit;
-	/**
-	 * 
-	 */
-	private Integer seckillSort;
+    /**
+     * id
+     */
+    @TableId
+    private Long id;
+    /**
+     * 该ID表示的是SessionID.
+     */
+    private Long promotionId;
+    /**
+     *
+     */
+    private Long promotionSessionId;
+    /**
+     *
+     */
+    private Long skuId;
+    /**
+     *
+     */
+    private BigDecimal seckillPrice;
+    /**
+     * 秒杀的数量.
+     */
+    private Integer seckillCount;
+    /**
+     * 没人限制秒杀的数量.
+     */
+    private BigDecimal seckillLimit;
+    /**
+     * 秒杀排序.
+     */
+    private Integer seckillSort;
 
 }
